Auth: added authentication & authorization to gateway and services. Added Auth service.

+++ user-service/src/main/java/pl/piomin/services/service/controller/UserController.java
@@ -8,12 +8,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 import pl.piomin.services.service.model.document.User;
 import pl.piomin.services.service.service.UserService;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 @RestController
@@ -21,31 +23,41 @@ public class UserController {
     private static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
     @Autowired private UserService userService;
 
-    @GetMapping("/{userId}")
+    @GetMapping("/get/{userId}")
     public ResponseEntity<User> getUserById(@PathVariable String userId) {
         LOGGER.info("Getting user " + userId);
         User foundUser = userService.getUser(userId);
-        return ResponseEntity.ok(foundUser);
+        if (foundUser != null) return new ResponseEntity<>(foundUser, HttpStatus.OK);
+        else return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
     @PostMapping("/create")
-    public ResponseEntity<User> createUser(@RequestBody User newUser) {
-        LOGGER.info("Creating user " + newUser.getDisplayName());
-        User createdUser = userService.createUser(newUser);
-        return ResponseEntity.ok(createdUser);
+    public ResponseEntity<User> createUser(@RequestBody User newUser,
+                @RequestHeader(value = "role") String requesterRole) {
+        if (requesterRole.equals("ADMIN")) {
+            User createdUser = userService.createUser(newUser);
+            return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
+        }
+        else return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping("/update/{userId}")
-    public ResponseEntity<User> updateUser(@RequestBody User newUser, @PathVariable String userId) {
-        LOGGER.info("Updating user " + userId);
-        User updatedUser = userService.updateUser(newUser, userId);
-        return ResponseEntity.ok(updatedUser);
+    public ResponseEntity<User> updateUser(@RequestBody User newUser, @PathVariable String userId,
+                @RequestHeader(value = "userId") String requesterId) {
+        if (requesterId.equals(userId)) {
+            User updatedUser = userService.updateUser(newUser, userId);
+            return new ResponseEntity<>(updatedUser, HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
     }
 
     @DeleteMapping("/delete/{userId}")
-    public ResponseEntity<String> deleteUser(@PathVariable String userId) {
-        LOGGER.info("Deleting user " + userId);
-        userService.deleteUser(userId);
-        return ResponseEntity.ok("User deleted.");
+    public ResponseEntity<String> deleteUser(@PathVariable String userId,
+                @RequestHeader(value = "role") String requesterRole) {
+        if (requesterRole.equals("ADMIN")) {
+            userService.deleteUser(userId);
+            return new ResponseEntity<>("User deleted.", HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
     }
 }
